We finally inplemented background guys!
